Extract txt report to own module
