+++ detekt-cli/src/main/kotlin/io/gitlab/arturbosch/detekt/cli/ReportPath.kt
@@ -1,8 +1,8 @@
 package io.gitlab.arturbosch.detekt.cli
 
 import io.github.detekt.report.html.HtmlOutputReport
+import io.github.detekt.report.txt.TxtOutputReport
 import io.github.detekt.report.xml.XmlOutputReport
-import io.gitlab.arturbosch.detekt.cli.out.TxtOutputReport
 import java.nio.file.Path
 import java.nio.file.Paths
