--- detekt-cli/src/main/resources/META-INF/services/io.gitlab.arturbosch.detekt.api.OutputReport
@@ -1,1 +1,0 @@
-io.gitlab.arturbosch.detekt.cli.out.TxtOutputReport
