+++ detekt-report-txt/src/main/resources/META-INF/services/io.gitlab.arturbosch.detekt.api.OutputReport
@@ -1,0 +1,1 @@
+io.github.detekt.report.txt.TxtOutputReport
