+++ settings.gradle.kts
@@ -10,6 +10,7 @@ include(
     "detekt-parser",
     "detekt-psi-utils",
     "detekt-report-html",
+    "detekt-report-txt",
     "detekt-report-xml",
     "detekt-rules",
     "detekt-sample-extensions",
